Debugging ICPC 2023 Mock P2
The logic for the solution was correct.

There was a logical error in the code.
In one of the loops, the array indexing was done using the incorrect variable, leading to the testcase passing even if the second condition wasn't met.

This was pointed out to me by a friend of mine, and resolved.

In its current state, the code past the online judge, and the problem stands solved.