Codeforces - Cut Ribbon
Started working on the cut ribbon problem from codeforces (1300 rated).
Came up with solution, wrote in form of algorithm.
Implementation in code pending.
Said implementation will take some time; using the concept of structs, which is new to me.

Deleted trial file, since there is no need for it to be in the repository.